# first-service/src/main/resources/swagger-OrganizationServiceOpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Organization service REST endpoints",
    "description": "These are an Organization service REST endpoints based on the OpenAPI 3.0 specification. SOA.",
    "contact": {
      "name": "developers",
      "email": "bordun.2002@mail.ru"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:9099/api/v1",
      "description": "Organization server"
    }
  ],
  "tags": [
    {
      "name": "basic set",
      "description": "Basic set of endpoints"
    },
    {
      "name": "advanced set",
      "description": "Advanced set of operations"
    }
  ],
  "paths": {
    "/organizations": {
      "post": {
        "tags": [
          "basic set"
        ],
        "summary": "",
        "operationId": "addOrganization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationWithoutId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns added Organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "get": {
        "tags": [
          "basic set"
        ],
        "summary": "Get Organizations with preferences",
        "description": "sorting and filtration fields: id, name, coordinates(x and y), creationDate, annualTurnover, type, officialAddress(street and zipCode)",
        "operationId": "getOrganizationsWithPreferences",
        "parameters": [
          {
            "name": "filters",
            "in": "query",
            "description": "Filters. Choose fields and comparison(==, !=, >, <, >=, <=)",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "id==1",
                "pattern": "^(id|name|coordinates\\.x|coordinates\\.y|creationDate|annualTurnover|type|officialAddress\\.street|officialAddress\\.zipCode)(==|!=|>|<|<=|>=).+$"
              }
            },
            "style": "deepObject"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Поле, по которому нужно сортировать (enum с множественным выбором).",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "name",
                  "coordinates.x",
                  "coordinates.y",
                  "creationDate",
                  "annualTurnover",
                  "type",
                  "officialAddress.street",
                  "officialAddress.zipCode"
                ]
              }
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "nullable": true,
              "format": "int32",
              "default": 1,
              "description": "Number of current page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "nullable": true,
              "format": "int32",
              "default": 5,
              "description": "Amount of Organizations per page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns Organization by ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/organizations/{id}": {
      "put": {
        "tags": [
          "basic set"
        ],
        "summary": "",
        "operationId": "updateOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID to update",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationWithoutId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns updated Organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "get": {
        "tags": [
          "basic set"
        ],
        "summary": "",
        "operationId": "getOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID to get",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns Organization by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "delete": {
        "tags": [
          "basic set"
        ],
        "summary": "",
        "operationId": "deleteOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID to delete",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/organizations/official-address/{official-address}": {
      "delete": {
        "tags": [
          "advanced set"
        ],
        "summary": "",
        "operationId": "deleteOrganizationByOfficialAddress",
        "parameters": [
          {
            "name": "official-address",
            "in": "path",
            "description": "Official address for Organization deleting",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Address"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/organizations/annual-turnover/sum": {
      "get": {
        "tags": [
          "advanced set"
        ],
        "summary": "Get sum of all Organizations annual turnovers",
        "operationId": "getSumOfOrganizationsAnnualTurnovers",
        "responses": {
          "200": {
            "description": "Returns sum of all Organizations annual turnovers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlyAnnualTurnover"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/organizations/annual-turnover/grouped": {
      "get": {
        "tags": [
          "advanced set"
        ],
        "summary": "",
        "operationId": "getGroupedOrganizationsByAnnualTurnover",
        "responses": {
          "200": {
            "description": "Returns grouped Organizations by annual turnover",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnualTurnoverOrganizationsCount"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "default": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Ok": {
        "description": "The operation was performed successfully",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "properties": {
                    "code": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Code"
                        },
                        {
                          "example": 200
                        }
                      ]
                    },
                    "message": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Message"
                        },
                        {
                          "example": "OK"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "NoContent": {
        "description": "No Content"
      },
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "properties": {
                    "code": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Code"
                        },
                        {
                          "example": 404
                        }
                      ]
                    },
                    "message": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Message"
                        },
                        {
                          "example": "Resource was not found"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "The request was bad composed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server error",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "properties": {
                    "code": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Code"
                        },
                        {
                          "example": 500
                        }
                      ]
                    },
                    "message": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Message"
                        },
                        {
                          "example": "Internal Server error"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "UnexpectedError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "properties": {
                    "code": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Code"
                        },
                        {
                          "example": "XXX"
                        }
                      ]
                    },
                    "message": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Message"
                        },
                        {
                          "example": "Unexpected error"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
      "Id": {
        "type": "integer",
        "format": "int32",
        "minimum": 1,
        "example": 1,
        "description": "Organization ID"
      },
      "AnnualTurnover": {
        "type": "number",
        "format": "double",
        "minimum": 0,
        "exclusiveMinimum": true,
        "example": 4.5,
        "description": "Organization annual turnover"
      },
      "OnlyAnnualTurnover": {
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/AnnualTurnover"
          }
        }
      },
      "EmployeesCount": {
        "type": "integer",
        "format": "int64",
        "minimum": 0,
        "example": 5,
        "description": "Employees count in organization"
      },
      "AnnualTurnoverOrganizationsCount": {
        "type": "object",
        "required": [
          "annualTurnover",
          "organizationsCount"
        ],
        "properties": {
          "annualTurnover": {
            "$ref": "#/components/schemas/AnnualTurnover"
          },
          "organizationsCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 5,
            "description": "Amount of Organizations with such annual turnover value"
          }
        }
      },
      "SortingOrFilteringField": {
        "type": "string",
        "enum": [
          "id",
          "name",
          "coordinates.x",
          "coordinates.y",
          "creationDate",
          "annualTurnover",
          "type",
          "officialAddress.street",
          "officialAddress.zipcode"
        ],
        "example": "id",
        "description": "Type of filtration"
      },
      "FiltrationPair": {
        "type": "object",
        "required": [
          "field",
          "value"
        ],
        "properties": {
          "field": {
            "$ref": "#/components/schemas/SortingOrFilteringField"
          },
          "value": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "example": 1,
            "description": "Field value"
          }
        }
      },
      "Organization": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrganizationWithoutId"
          },
          {
            "type": "object"
          },
          {
            "required": [
              "id"
            ]
          },
          {
            "properties": {
              "id": {
                "$ref": "#/components/schemas/Id"
              }
            }
          }
        ]
      },
      "OrganizationsPreferences": {
        "type": "object",
        "required": [
          "filtrationFields",
          "sortingFields"
        ],
        "properties": {
          "filtrationFields": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FiltrationPair"
            }
          },
          "sortingFields": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/SortingOrFilteringField"
            }
          }
        }
      },
      "OrganizationWithoutId": {
        "type": "object",
        "required": [
          "name",
          "coordinates",
          "creationDate",
          "annualTurnover",
          "type",
          "officialAddress"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "OAO Aviasales",
            "description": "Organization name"
          },
          "coordinates": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "creationDate": {
            "type": "string",
            "format": "date",
            "example": "2024-03-12T00:00:00.000Z",
            "description": "Organization creation date, generates automatically"
          },
          "annualTurnover": {
            "$ref": "#/components/schemas/AnnualTurnover"
          },
          "type": {
            "allOf": [
              {
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/OrganizationType"
              },
              {
                "nullable": true
              }
            ]
          },
          "officialAddress": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "nullable": true
              }
            ]
          }
        }
      },
      "Coordinates": {
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "properties": {
          "x": {
            "type": "number",
            "format": "double",
            "example": 4.9,
            "description": "x coordinate"
          },
          "y": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "example": 6,
            "description": "y coordinate"
          }
        }
      },
      "Address": {
        "type": "object",
        "required": [
          "street",
          "zipCode"
        ],
        "properties": {
          "street": {
            "type": "string",
            "minLength": 1,
            "example": "Kosovo street",
            "description": "Address street"
          },
          "zipCode": {
            "type": "string",
            "nullable": true,
            "example": "AE84C",
            "description": "Address zipcode"
          }
        }
      },
      "OrganizationType": {
        "type": "string",
        "enum": [
          "COMMERCIAL",
          "PUBLIC",
          "OPEN_JOINT_STOCK_COMPANY"
        ],
        "example": "COMMERCIAL",
        "description": "Type of Organization"
      },
      "Error": {
        "type": "object",
        "required": [
          "httpCode",
          "message",
          "time"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/Code"
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "time": {
            "$ref": "#/components/schemas/Time"
          }
        }
      },
      "Code": {
        "type": "string",
        "format": "int32",
        "example": 400,
        "description": "Error code"
      },
      "Message": {
        "type": "string",
        "example": "Bad Request",
        "description": "Error message"
      },
      "Time": {
        "type": "string",
        "format": "date-time",
        "example": "2022-09-19T00:00:09.000Z",
        "description": "Error time"
      }
    }
  }
}